refactor(event-handlers): move event-handlers into each domain
