+++ src/event-handlers/deposit/deposit-event.handler.spec.ts
@@ -1,13 +1,13 @@
 import { Test, TestingModule } from '@nestjs/testing';
 import { DepositEventHandler } from './deposit-event.handler';
-import { DepositMatchedEvent } from '../features/deposit/domain/events/deposit-matched.event';
+import { DepositMatchedEvent } from 'src/features/deposit/domain/events/deposit-matched.event';
 import { NotificationService } from 'src/features/notification/notification.service';
 import { User } from 'src/entities/user.entity';
 import { Funding } from 'src/entities/funding.entity';
 import { Donation } from 'src/entities/donation.entity';
 import { Notification } from 'src/entities/notification.entity';
-import { Deposit } from '../entities/deposit.entity';
-import { ProvisionalDonation } from '../entities/provisional-donation.entity';
+import { Deposit } from 'src/entities/deposit.entity';
+import { ProvisionalDonation } from 'src/entities/provisional-donation.entity';
 import { ImageType } from 'src/enums/image-type.enum';
 import { AuthType } from 'src/enums/auth-type.enum';
 import { FundTheme } from 'src/enums/fund-theme.enum';
@@ -20,14 +20,13 @@ import { Gift } from 'src/entities/gift.entity';
 import { CreateDonationUseCase } from 'src/features/donation/commands/create-donation.usecase';
 import { IncreaseFundSumUseCase } from 'src/features/funding/commands/increase-fundsum.usecase';
 import { GetDonationsByFundingUseCase } from 'src/features/donation/queries/get-donations-by-funding.usecase';
-import { createMockProvider } from '../tests/create-mock-repository';
-import { DepositUnmatchedEvent } from '../features/deposit/domain/events/deposit-unmatched.event';
-import { CreateNotificationDto } from 'src/features/notification/dto/create-notification.dto';
+import { createMockProvider } from 'src/tests/create-mock-repository';
+import { DepositUnmatchedEvent } from 'src/features/deposit/domain/events/deposit-unmatched.event';
 import { NotiType } from 'src/enums/noti-type.enum';
 import { DecreaseFundSumUseCase } from 'src/features/funding/commands/decrease-fundsum.usecase';
 import { FindAllAdminsUseCase } from 'src/features/admin/queries/find-all-admins.usecase';
 import { CreateDonationCommand } from 'src/features/donation/commands/create-donation.command';
-import { DepositPartiallyMatchedEvent } from '../features/deposit/domain/events/deposit-partially-matched.event';
+import { DepositPartiallyMatchedEvent } from 'src/features/deposit/domain/events/deposit-partially-matched.event';
 import { EventEmitter2 } from '@nestjs/event-emitter';
 import { DepositStatus } from 'src/enums/deposit-status.enum';
