+++ src/event-handlers/donation/donation-event-handler.integration.spec.ts
@@ -2,30 +2,29 @@ import { Test, TestingModule } from '@nestjs/testing';
 import { DonationEventHandler } from './donation-event-handler';
 import { TypeOrmModule } from '@nestjs/typeorm';
 import { EventEmitterModule } from '@nestjs/event-emitter';
-import { NotificationService } from '../features/notification/notification.service';
-import { FindAllAdminsUseCase } from '../features/admin/queries/find-all-admins.usecase';
-import { DeleteDepositUseCase } from '../features/deposit/commands/delete-deposit.usecase';
-import { DecreaseFundSumUseCase } from '../features/funding/commands/decrease-fundsum.usecase';
-import { DonationRefundRequestedEvent } from '../features/donation/domain/events/donation-refund-requested.event';
-import { DonationRefundCancelledEvent } from '../features/donation/domain/events/donation-refund-cancelled.event';
-import { DonationDeletedEvent } from '../features/donation/domain/events/donation-deleted.event';
-import { User } from '../entities/user.entity';
-import { Notification } from '../entities/notification.entity';
-import { Funding } from '../entities/funding.entity';
-import { Donation } from '../entities/donation.entity';
-import { Deposit } from '../entities/deposit.entity';
+import { NotificationService } from '../../features/notification/notification.service';
+import { FindAllAdminsUseCase } from '../../features/admin/queries/find-all-admins.usecase';
+import { DeleteDepositUseCase } from '../../features/deposit/commands/delete-deposit.usecase';
+import { DecreaseFundSumUseCase } from '../../features/funding/commands/decrease-fundsum.usecase';
+import { DonationRefundRequestedEvent } from '../../features/donation/domain/events/donation-refund-requested.event';
+import { DonationRefundCancelledEvent } from '../../features/donation/domain/events/donation-refund-cancelled.event';
+import { User } from '../../entities/user.entity';
+import { Notification } from '../../entities/notification.entity';
+import { Funding } from '../../entities/funding.entity';
+import { Donation } from '../../entities/donation.entity';
+import { Deposit } from '../../entities/deposit.entity';
 import { DataSource, Repository } from 'typeorm';
-import { GiftogetherExceptions } from '../filters/giftogether-exception';
-import { DepositFsmService } from '../features/deposit/domain/deposit-fsm.service';
+import { GiftogetherExceptions } from '../../filters/giftogether-exception';
+import { DepositFsmService } from '../../features/deposit/domain/deposit-fsm.service';
 import { getRepositoryToken } from '@nestjs/typeorm';
 import {
   createMockUser,
   createMockDeposit,
   createMockFunding,
   createMockDonation,
-} from '../tests/mock-factory';
-import { createDataSourceOptions } from '../tests/data-source-options';
-import entities from '../entities/entities';
+} from '../../tests/mock-factory';
+import { createDataSourceOptions } from '../../tests/data-source-options';
+import entities from '../../entities/entities';
 
 describe('DonationEventHandler (Integration)', () => {
   let module: TestingModule;
