+++ src/event-handlers/provisional-donation/provisional-donation-event-handler.ts
@@ -1,13 +1,13 @@
 import { Injectable } from '@nestjs/common';
-import { ProvisionalDonationFsmService } from '../features/donation/domain/services/provisional-donation-fsm.service';
+import { ProvisionalDonationFsmService } from 'src/features/donation/domain/services/provisional-donation-fsm.service';
 import { EventEmitter2, OnEvent } from '@nestjs/event-emitter';
 import { InjectRepository } from '@nestjs/typeorm';
 import { ProvisionalDonation } from 'src/entities/provisional-donation.entity';
 import { Repository } from 'typeorm';
 import { DepositMatchedEvent } from 'src/features/deposit/domain/events/deposit-matched.event';
 import { DepositPartiallyMatchedEvent } from 'src/features/deposit/domain/events/deposit-partially-matched.event';
-import { ProvisionalDonationApprovedEvent } from '../features/donation/domain/events/provisional-donation-approved.event';
-import { ProvisionalDonationPartiallyMatchedEvent } from '../features/donation/domain/events/provisional-donation-partially-matched.event';
+import { ProvisionalDonationApprovedEvent } from 'src/features/donation/domain/events/provisional-donation-approved.event';
+import { ProvisionalDonationPartiallyMatchedEvent } from 'src/features/donation/domain/events/provisional-donation-partially-matched.event';
 
 @Injectable()
 export class ProvisionalDonationEventHandler {
